Data Visualization of Bird Strikes between 2000 – 2011
Problem Statement:-
Transport and communication are in the crucial domain in the field of analytics.
Environmental impacts and safety are, nowadays, two major concerns of the scientific
community with respect to transport scenarios and to the ever-growing urban areas.
These issues gain more importance due to the increasing amount of vehicles and
people. Seeking new solutions is reaching a point where available technologies and
artificial intelligence, especially MAS, are being recognized as ways to cope with and
tackle these kinds of problems in a distributed and more appropriate way.
A bird strike is strictly defined as a collision between a bird and an aircraft which is in
flight or on a take-off or landing roll. The term is often expanded to cover other wildlife
strikes - with bats or ground animals. Bird Strike is common and can be a significant
threat to aircraft safety. For smaller aircraft, significant damage may be caused to the
aircraft structure and all aircraft, especially jet-engine ones, are vulnerable to the loss
of thrust which can follow the ingestion of birds into engine air intakes. This has
resulted in several fatal accidents.
Bird strikes may occur during any phase of flight, but are most likely during the
take-off, initial climb, approach and landing phases due to the greater numbers of birds
in flight at lower levels. To have a closer look the following document visually depicts
the data collected on Bird Strikes by FAA between 2000-2011